# docs/swagger.yaml
basePath: /api
definitions:
  LoginInput:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  LoginResponse:
    properties:
      access_token:
        type: string
      expired_date:
        type: string
      refresh_token:
        type: string
      type:
        type: string
    type: object
  RegisterInput:
    properties:
      fio:
        type: string
      iin:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - fio
    - iin
    - password
    - phone
    type: object
...
  FileResponse:
    type: object
    properties:
      url:
        type: string
      uploaded_at:
        type: string
        format: datetime
  UserResponse:
    properties:
      fio:
        type: string
      id:
        type: integer
      iin:
        type: string
      phone:
        type: string
      type:
        type: string
      birthday:
        type: string
        format: datetime
      avatar:
        $ref: '#/definitions/FileResponse'
      wallet:
        type: object
        properties:
          balance:
            type: number
          bonus:
            type: number
    type: object
...
host: api.paydal.kz
schemes:
  - "https"
info:

  title: Paydal API документация
  version: "1.0"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      Передайте полученный токен при авторизаций "Bearer {token}".
security:
  - bearerAuth: []
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Логин
      parameters:
      - description: Login Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponse'
      summary: Логин с использованием phone и password
      tags:
      - Авторизация
...
  /p/user:
    get:
      security:
        - Bearer: [ ]
      consumes:
      - application/json
      description: Получить пользотваля
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
              data:
                $ref: '#/definitions/UserResponse'
      summary:  Получить пользотваля
      tags:
      - Пользователь
...
  /register:
    post:
      consumes:
      - application/json
      description: Регистрация  нового пользователя
      parameters:
      - description: Register Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
      summary: Регистрация
      tags:
      - Авторизация
...
swagger: "2.0"